auto confirm School Contribution. add system claim

+++ src/Core/Data/Dto/School/CreateSchoolImageRequestDto.cs
@@ -8,7 +8,7 @@ namespace GamaEdtech.Data.Dto.School
     {
         public required IFormFile File { get; set; }
         public required FileType FileType { get; set; }
-        public int SchoolId { get; set; }
+        public long SchoolId { get; set; }
         public DateTimeOffset CreationDate { get; set; }
         public int CreationUserId { get; set; }
     }
