+++ src/Domain/Entity/ISchoolId.cs
@@ -2,6 +2,6 @@ namespace GamaEdtech.Domain.Entity
 {
     public interface ISchoolId
     {
-        int SchoolId { get; set; }
+        long SchoolId { get; set; }
     }
 }
